split crate into lib.rs and main.rs for testing purposes
